Factor out usePermalinkEvent
